backend/requirements.txt
# FastAPI & Web Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Stable Diffusion & AI
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.1
omegaconf>=2.3.0

# Image Processing
Pillow>=10.2.0
opencv-python>=4.9.0
imageio>=2.33.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
